Migrate robolectric tests
