+++ gnd/build.gradle
@@ -93,6 +93,9 @@ android {
             includeAndroidResources = true
         }
     }
+    hilt {
+        enableTransformForLocalTests = true
+    }
 }
 
 
@@ -209,6 +212,8 @@ dependencies {
     testImplementation 'androidx.test:core:1.2.0'
     testImplementation 'org.robolectric:robolectric:4.3.1'
     testImplementation 'android.arch.core:core-testing:1.1.1'
+    testImplementation 'com.google.dagger:hilt-android-testing:2.28-alpha'
+    testAnnotationProcessor 'com.google.dagger:hilt-android-compiler:2.28-alpha'
 
     androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0-alpha4', {
         exclude group: 'com.android.support', module: 'support-annotations'
